+++ keyboards/yushakobo/navpad/10/rules.mk
@@ -7,7 +7,7 @@ BOOTLOADER = caterina
 # Build Options
 #   change yes to no to disable
 #
-BOOTMAGIC_ENABLE = lite     # Virtual DIP switch configuration
+BOOTMAGIC_ENABLE = yes      # Virtual DIP switch configuration
 MOUSEKEY_ENABLE = yes       # Mouse keys
 EXTRAKEY_ENABLE = yes       # Audio control and System control
 CONSOLE_ENABLE = no         # Console for debug
@@ -23,4 +23,5 @@ AUDIO_ENABLE = no           # Audio output
 
 ENCODER_ENABLE = yes
 
+SRC += navpad_prefs.c
 DEFAULT_FOLDER = yushakobo/navpad/10/rev1
