+++ keyboards/yushakobo/navpad/10_helix_r/10_helix_r.c
@@ -16,21 +16,6 @@
 
 #include "10_helix_r.h"
 
-bool process_record_kb(uint16_t keycode, keyrecord_t *record) {
-  switch (keycode) {
-  case TAP_00:
-    if (record->event.pressed){
-      tap_code(KC_P0);
-      tap_code(KC_P0);
-      }
-    break;
-  
-  default:
-    break;
-  }
-  return true; 
-}
-
 #ifdef RGBLIGHT_LAYERS
 
 #define LOCK_COLOR_1 HSV_TEAL
@@ -67,4 +52,53 @@ void keyboard_post_init_kb(void) {
     rgblight_layers = navpad_10_rgb_layers;
 }
 
+#endif
+
+#ifdef ENCODER_ENABLE
+bool encoder_update_kb(uint8_t index, bool clockwise) {
+  if (index == 0) { /* Navpad side encoder */
+    if (IS_LAYER_ON(_BASE)){
+      if (clockwise) {
+        tap_code16(KC_VOLU);
+      } else {
+        tap_code16(KC_VOLD);
+      }
+    } else if (IS_LAYER_ON(_FN2)){
+      if (clockwise) {
+        rgblight_increase_sat();
+      } else {
+        rgblight_decrease_sat();
+      }
+    } else if (IS_LAYER_ON(_FN1)){
+      if (clockwise) {
+        rgblight_increase_hue();
+      } else {
+        rgblight_decrease_hue();
+      }
+    }
+  }
+  if (index == 1) { /* Helix side encoder */
+    if (IS_LAYER_ON(_BASE)){
+      if (clockwise) {
+        tap_code16(KC_VOLU);
+      } else {
+        tap_code16(KC_VOLD);
+      }
+    } else if (IS_LAYER_ON(_FN2)){
+      if (clockwise) {
+        rgblight_increase_sat();
+      } else {
+        rgblight_decrease_sat();
+      }
+    } else if (IS_LAYER_ON(_FN1)){
+      if (clockwise) {
+        rgblight_increase_hue();
+      } else {
+        rgblight_decrease_hue();
+      }
+    }
+
+  }
+  return false;
+}
 #endif
