+++ keyboards/yushakobo/navpad/10_helix_r/rules.mk
@@ -23,3 +23,5 @@ AUDIO_ENABLE = no           # Audio output
 
 SPLIT_KEYBOARD = yes
 ENCODER_ENABLE = yes
+
+SRC += navpad_prefs.c
